Created uiview subclass with an api designed for border color/width and corner radius
